# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "summary": "Registrar novo usuário",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "summary": "Autenticar usuário com e-mail e senha",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido ou expirado"
          }
        },
        "summary": "Renovar sessão com refresh token (rotação)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token JWT inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retornar usuário autenticado",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token JWT inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Encerrar sessão e invalidar tokens ativos",
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "API Financeiro",
    "description": "API para gerenciamento financeiro pessoal",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Marcos Silva",
            "description": "Nome completo do usuário"
          },
          "email": {
            "type": "string",
            "example": "marcos@email.com",
            "description": "E-mail único do usuário"
          },
          "password": {
            "type": "string",
            "example": "Forte@123",
            "description": "Senha forte: mínimo 8 caracteres, com maiúscula, minúscula, número e símbolo"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Marcos Silva"
          },
          "email": {
            "type": "string",
            "example": "marcos@email.com"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "AuthSessionDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access.payload.signature"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh.payload.signature"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "example": 3600
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "marcos@email.com",
            "description": "E-mail de acesso"
          },
          "password": {
            "type": "string",
            "example": "Forte@123",
            "description": "Senha de acesso"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token JWT válido emitido no login ou registro",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh.payload.signature"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Sessão encerrada com sucesso"
          }
        },
        "required": [
          "message"
        ]
      }
    }
  }
}